requirements.txt
openai>=1.0.0
httpx>=0.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tqdm>=4.65.0
pillow>=9.5.0
